chore(site): move collection page to app dir
