+++ app/collection/components/collection-item-metadata/types.d.ts
