+++ app/collection/components/collection-item/types.d.ts
