+++ app/collection/components/collection-list/styles.module.scss
