+++ app/collection/components/collection-list/types.d.ts
