+++ app/collection/components/collection-search/styles.module.scss
