+++ app/collection/components/collection-search/types.d.ts
@@ -1,6 +1,6 @@
 import { ChangeEventHandler } from "react"
 
-export type SearchProps = {
+export type CollectionSearchProps = {
   value: string
   handleChange: ChangeEventHandler<HTMLInputElement>
 }
