--- components/collection/index.ts
@@ -1,3 +1,0 @@
-export * from "./collection-list"
-export * from "./collection-filter"
-export * from "./search"
